+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - upgrade reflex to 0.8.14
+- updated template to https://github.com/robert-koch-institut/mex-template/commit/33ca3a
 - bump python to 3.13
 - updated template to https://github.com/robert-koch-institut/mex-template/commit/ad26e4
 - bump common to 1.11